Feeds: isolate the util test from the command test
